perf: split web studio heavy chunks
